Issue #4 - Renaming Sampled* classes to *History
