+++ management-model/src/main/java/org/terracotta/management/stats/history/AverageHistory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terracotta.management.stats.sampled;
+package org.terracotta.management.stats.history;
 
 import org.terracotta.management.stats.AbstractMeasurableStatistic;
 import org.terracotta.management.stats.Sample;
@@ -22,10 +22,10 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 /**
- * @author Ludovic Orban
+ * @author Mathieu Carbou
  */
-public class SampledDuration extends AbstractMeasurableStatistic<List<Sample<Long>>, TimeUnit> {
-  public SampledDuration(String name, List<Sample<Long>> samples, TimeUnit timeUnit) {
+public class AverageHistory extends AbstractMeasurableStatistic<List<Sample<Double>>, TimeUnit> {
+  public AverageHistory(String name, List<Sample<Double>> samples, TimeUnit timeUnit) {
     super(name, samples, timeUnit);
   }
 }
